fix(audio IN-3): throw on compact-XWB dataLength underflow instead of wrapping
The compact-format entry length computation (gap to next entry's offset
minus deviation, or remainder of the wave-data segment for the last entry)
was plain uint32_t subtraction with no guard, so a corrupt/adversarial file
could underflow it to a huge value near UINT32_MAX. The downstream bounds
check in WaveBank::GetSoundEffect summed dataOffset+dataLength in uint32_t
too, so that huge length could itself wrap the check back into range and
cause a heap over-read.

Both underflow sites in ParseXwb now check in uint64_t and throw
std::runtime_error on underflow (per plan_audio.md D7: throw rather than
silently clamp), and the WaveBank bounds check now sums in uint64_t so it
can no longer wrap.

+++ src/Microsoft/Xna/Framework/Audio/WaveBank.cpp
@@ -217,9 +217,10 @@ namespace Microsoft::Xna::Framework::Audio
 
         const auto& entry = xactImpl_->data.entries[waveIndex];
 
-        // Check that data range is valid
+        // Check that data range is valid. Widen to 64-bit before summing so a corrupt/adversarial
+        // entry can't wrap this check via uint32_t overflow and pass with an out-of-range offset.
         const auto& fd = xactImpl_->data.fileData;
-        if (entry.dataOffset + entry.dataLength > fd.size())
+        if (static_cast<uint64_t>(entry.dataOffset) + entry.dataLength > fd.size())
         {
             std::cerr << "[WaveBank] Wave " << waveIndex << " data out of range\n";
             return nullptr;
